Adding in posts section
So far I’ve just set this as a separate paragraph. It’s not actively
going to refresh like on the Missourian site (because I don’t know how
we’d even do that.) Not sure if this should be a separate div. Comment
if there needs to be changes!